feat(spadl): kloppy Sportec + Metrica + _SoccerActionCoordinateSystem fix — silly-kicks 1.6.0 (#11)
Adds Provider.SPORTEC and Provider.METRICA to the kloppy converter's
_SUPPORTED_PROVIDERS whitelist. Empirical verification on real kloppy
fixture data confirms zero new event-type mappings are required — both
providers' serializers emit only event types already covered by
_MAPPED_EVENT_TYPES ∪ _EXCLUDED_EVENT_TYPES.

Fixes a latent bug in _SoccerActionCoordinateSystem (no __init__, but
convert_to_actions instantiates it with pitch_length=/pitch_width=
kwargs) that has prevented the kloppy converter from running end-to-end
on any real EventDataset since silly-kicks 1.0.0. The pre-existing
mock-based tests never exercised dataset.transform().

Aligns the kloppy converter with the established silly-kicks convention
of clamping output coords to [0, field_length] × [0, field_width],
matching StatsBomb / Wyscout / Opta. Metrica events emit slight off-pitch
coords (x ∈ [-1.62, 104.63] empirically) within source-recording-noise
tolerance.

Also fixes 2 pyright errors in silly_kicks/xthreat.py:402 surfaced by
newer pandas-stubs / numpy-stubs versions (explicit dtype=np.float64
on two np.linspace calls).

13 new test functions (16 parametrized cases) under tests/spadl/test_kloppy.py
including a regression test for the _SoccerActionCoordinateSystem bug,
real-fixture end-to-end suites for Sportec and Metrica vendored from
kloppy's BSD-3-licensed test files, a parametrized coordinate-clamping
test, and a per-provider ConversionReport invariant test.

C4 architecture diagram (docs/c4/architecture.{dsl,html}) regenerated
to reflect the expanded kloppy gateway coverage and clamping convention.

+++ silly_kicks/xthreat.py
@@ -397,8 +397,8 @@ class ExpectedThreat:
             interp = self.interpolator()
             l = int(spadlconfig.field_length * 10)
             w = int(spadlconfig.field_width * 10)
-            xs = np.linspace(0, spadlconfig.field_length, l)
-            ys = np.linspace(0, spadlconfig.field_width, w)
+            xs = np.linspace(0, spadlconfig.field_length, l, dtype=np.float64)
+            ys = np.linspace(0, spadlconfig.field_width, w, dtype=np.float64)
             grid = interp(xs, ys)
 
         ratings = np.empty(len(actions))
